+++ src/components/Dialog/ExportDialog/index.ts
@@ -1,0 +1,2 @@
+export * from "./ExportDialog";
+export type * from "./types";