+++ src/components/Dialog/ExportDialog/types.ts
@@ -1,0 +1,11 @@
+import { ReactNode } from "react";
+
+// components
+import { ButtonPropsType, DialogPropsType } from "components";
+
+export interface ExportDialogPropsType extends DialogPropsType {
+  handleSubmit: () => void;
+  isLoading?: boolean;
+  extraFields?: ReactNode;
+  extraActions?: ButtonPropsType[];
+}